tweaked layout, borrower can now manually mark items as returned

+++ grails-app/services/com/friendlibrary/MessageService.groovy
@@ -255,28 +255,41 @@ class MessageService {
 
   /*
   // Un-mark an item as loaned out, and inform the owner
+  // Can take message (i.e. from profile page) or user/item id's (i.e. from library page)
   */
   void confirmReturnRequest(params){
-    def message = Message.get(params.messageId)
-    assert message != null
-    def requestedItem = message.item
-    assert message != null
+    def message
+    def requestedItem
+    def requestingUser
+    def requestedUser
+
+    if(params.messageId != null){
+      message = Message.get(params.messageId)
+      requestedItem = message.item
+      requestingUser = message.sentTo
+      requestedUser = message.sentFrom
+      message.delete()
+    }else{
+      requestedItem = Item.get(params.requestedMedia)
+      requestingUser = User.get(params.requestingUser)
+      requestedUser = User.get(params.requestedUser)
+    }
     
     def confirmMessage = new Message(
-      sentFrom:message.sentTo,
-      sentTo:message.sentFrom,
-      body:"${message.sentTo.username} has returned the ${requestedItem.mediaType} \"${requestedItem.title}\" ",
+      sentFrom:requestingUser,
+      sentTo:requestedUser,
+      body:"${requestingUser.username} has returned the ${requestedItem.mediaType} \"${requestedItem.title}\" ",
       type:"Item Return Confirm"
     )
+
     confirmMessage.save(failOnError:true)
-    message.sentFrom.addToInMessages(confirmMessage)
-    message.sentTo.addToOutMessages(confirmMessage)
+    requestedUser.addToInMessages(confirmMessage)
+    requestingUser.addToOutMessages(confirmMessage)
 
     requestedItem.loanedOut = false
     requestedItem.loanedTo = null
     requestedItem.save(failOnError:true)
     requestedItem.library.available[requestedItem.mediaType + "s"] += 1
-    message.delete()
   }
 
   /*
